Move some functions around.
